requirements.txt
streamlit
firebase-admin
python-dotenv
opencv-python
mediapipe
numpy
matplotlib
tk
absl-py